Update/add/set methods for solutions in OptimizationResult (#293)
* update/add/set methods for solutions in OptimizationResult

* tests ok

* ok now
