+++ src/Algorithm/Algorithm.jl
@@ -8,15 +8,12 @@ import ..Coluna
 using ..Containers
 using ..MathProg
 
-# To be deleted :
-import .MathProg: getsense, optimize! # because of branch
-
 # TO be deleted ???
 import .MathProg: FeasibilityStatus, TerminationStatus, AbstractStorage, EmptyStorage, getstorage, setprimalbound!, setdualbound!
 import .MathProg: OPTIMAL, TIME_LIMIT, NODE_LIMIT, OTHER_LIMIT, EMPTY_RESULT, NOT_YET_DETERMINED
 import ..Coluna: AbstractSense
 
-import .MathProg: get_lp_primal_sol, get_lp_dual_bound, setfeasibilitystatus!, setterminationstatus!, getvalue
+import .MathProg: getvalue
 
 using Logging
 using Printf
