+++ src/Algorithm/benders.jl
@@ -490,7 +490,7 @@ function bend_cutting_plane_main_loop!(
             return 
         end
 
-        add_lp_dual_sol!(algdata.incumbents, master_dual_sol, true)
+        set_lp_dual_sol!(algdata.incumbents, master_dual_sol)
         dual_bound = get_lp_dual_bound(algdata.incumbents)
         set_lp_dual_bound!(algdata.incumbents, dual_bound)
         set_ip_dual_bound!(algdata.incumbents, dual_bound)
@@ -515,7 +515,7 @@ function bend_cutting_plane_main_loop!(
             end
 
             # TODO: update bendcutgen stabilization
-            add_lp_primal_sol!(algdata.incumbents, master_primal_sol, true)
+            set_lp_primal_sol!(algdata.incumbents, master_primal_sol)
             set_lp_primal_bound!(algdata.incumbents, primal_bound)
             cur_gap = gap(primal_bound, dual_bound)
