+++ src/MathProg/optimizationresults.jl
@@ -122,9 +122,36 @@ mutable struct OptimizationResult{F<:AbstractFormulation,S<:Coluna.AbstractSense
     termination_status::TerminationStatus
     feasibility_status::FeasibilityStatus
     incumbents::ObjValues{S}
+    max_length_ip_primal_sols::Int
+    max_length_lp_primal_sols::Int
+    max_length_lp_dual_sols::Int
+    insert_function_ip_primal_sols::Function
+    insert_function_lp_primal_sols::Function
+    insert_function_lp_dual_sols::Function
     ip_primal_sols::Union{Nothing, Vector{PrimalSolution{F}}}
     lp_primal_sols::Union{Nothing, Vector{PrimalSolution{F}}}
     lp_dual_sols::Union{Nothing, Vector{DualSolution{F}}}
+end
+
+_sort!(sols::Vector{PrimalSolution{F}}, f::Function) where {F} = sort!(sols, by = x -> f(PrimalBound(x.model, x.value)))
+_sort!(sols::Vector{DualSolution{F}}, f::Function) where {F} = sort!(sols, by = x -> f(DualBound(x.model, x.value)))
+
+function bestbound(solutions::Vector{Sol}, max_len::Int, new_sol::Sol) where {Sol<:Coluna.Containers.Solution}
+    if length(solutions) < max_len
+        push!(solutions, new_sol)
+    else
+        solutions[end] = new_sol
+    end
+    _sort!(solutions, valueinminsense)
+    return
+end
+
+function pushfirst(solutions::Vector{Sol}, max_len::Int, new_sol::Sol) where {Sol<:Coluna.Containers.Solution}
+    pushfirst!(solutions, new_sol)
+    if length(solutions) > max_len
+        pop!(solutions)
+    end
+    return
 end
 
 """
@@ -135,14 +162,6 @@ Structure to be returned by all Coluna `optimize!` methods.
 
 Builds an empty OptimizationResult.
 """
-function OptimizationResult(model::M) where {M<:Coluna.Containers.AbstractModel}
-    S = getobjsense(model)
-    return OptimizationResult{M,S}(
-        NOT_YET_DETERMINED, UNKNOWN_FEASIBILITY, ObjValues(model), 
-        nothing, nothing, nothing
-    )
-end
-
 function OptimizationResult(
     form::F;
     feasibility_status::FeasibilityStatus = UNKNOWN_FEASIBILITY,
@@ -150,7 +169,13 @@ function OptimizationResult(
     ip_primal_bound = nothing,
     ip_dual_bound = nothing,
     lp_primal_bound = nothing,
-    lp_dual_bound = nothing
+    lp_dual_bound = nothing,
+    max_length_ip_primal_sols = 1,
+    max_length_lp_primal_sols = 1,
+    max_length_lp_dual_sols = 1,
+    insert_function_ip_primal_sols = bestbound,
+    insert_function_lp_primal_sols = bestbound,
+    insert_function_lp_dual_sols = bestbound
 ) where {F <: AbstractFormulation}
     incumbents = ObjValues(
         form;
@@ -161,7 +186,15 @@ function OptimizationResult(
     )
     S = getobjsense(form)
     result = OptimizationResult{F,S}(
-        termination_status, feasibility_status, incumbents,
+        termination_status, 
+        feasibility_status, 
+        incumbents,
+        max_length_ip_primal_sols,
+        max_length_lp_primal_sols,
+        max_length_lp_dual_sols,
+        insert_function_ip_primal_sols,
+        insert_function_lp_primal_sols,
+        insert_function_lp_dual_sols,
         nothing, nothing, nothing
     )
     return result
@@ -204,17 +237,17 @@ get_lp_primal_bound(res::OptimizationResult) = get_lp_primal_bound(res.incumbent
 get_ip_dual_bound(res::OptimizationResult) = get_ip_dual_bound(res.incumbents)
 get_lp_dual_bound(res::OptimizationResult) = get_lp_dual_bound(res.incumbents)
 
-ip_gap(res::OptimizationResult) = ip_gap(res.incumbents)
+update_ip_primal_bound!(res::OptimizationResult, val) = update_ip_primal_bound!(res.incumbents, val)
+update_ip_dual_bound!(res::OptimizationResult, val) = update_ip_dual_bound!(res.incumbents, val)
+update_lp_primal_bound!(res::OptimizationResult, val) = update_lp_primal_bound!(res.incumbents, val)
+update_lp_dual_bound!(res::OptimizationResult, val) = update_lp_dual_bound!(res.incumbents, val)
 
 set_ip_primal_bound!(res::OptimizationResult, val) = set_ip_primal_bound!(res.incumbents, val)
 set_lp_primal_bound!(res::OptimizationResult, val) = set_lp_primal_bound!(res.incumbents, val)
 set_ip_dual_bound!(res::OptimizationResult, val) = set_ip_dual_bound!(res.incumbents, val)
 set_lp_dual_bound!(res::OptimizationResult, val) = set_lp_dual_bound!(res.incumbents, val)
 
-update_ip_primal_bound!(res::OptimizationResult, val) = update_ip_primal_bound!(res.incumbents, val)
-update_ip_dual_bound!(res::OptimizationResult, val) = update_ip_dual_bound!(res.incumbents, val)
-update_lp_primal_bound!(res::OptimizationResult, val) = update_lp_primal_bound!(res.incumbents, val)
-update_lp_dual_bound!(res::OptimizationResult, val) = update_lp_dual_bound!(res.incumbents, val)
+ip_gap(res::OptimizationResult) = ip_gap(res.incumbents)
 
 function nb_ip_primal_sols(res::OptimizationResult)
     return res.ip_primal_sols === nothing ? 0 : length(res.ip_primal_sols)
@@ -255,51 +288,65 @@ function get_best_lp_dual_sol(res::OptimizationResult)
     return get_lp_dual_sols(res)[1]
 end
 
-# TODO : create set_ip_primal_sol if length list = 1
-function add_ip_primal_sol!(res::OptimizationResult{F,S}, solution::PrimalSolution{F}, force = false) where {F,S}
-    if res.ip_primal_sols === nothing
-        res.ip_primal_sols = PrimalSolution{F}[]
+# update : add the solution in the list iff its value is better than the incumbent
+# add : Add the solution in the list, update incumbents if necessary
+# set : add the solution, incumbent is not updated
+macro gen_new_sol_method(expr)
+    action_kw, sol_field, space_type = expr.args
+
+    method_name = Symbol(string(action_kw,"_",sol_field,"_sol!"))
+    max_len = Expr(:call, :getfield, :res, :(Symbol($(string("max_length_",sol_field,"_sols")))))
+    bound_type = Expr(:curly, Symbol(string(space_type,"Bound")), :S)
+    bound = Expr(:call, bound_type, :(getvalue(sol)))
+    update_func = Symbol(string("update_",sol_field,"_bound!"))
+    sol_type = Expr(:curly, Symbol(string(space_type,"Solution")), :F)
+    insert_method_name = Expr(:call, :getfield, :res, :(Symbol($(string("insert_function_",sol_field,"_sols")))))
+    field = Expr(:call, :getfield, :res, :(Symbol($(string(sol_field,"_sols")))))
+
+    body = Expr(:call, insert_method_name, field, max_len, :sol)
+    call_to_update = Expr(:call, update_func, :(res.incumbents), :b)
+    define_array = Expr(:call, :setfield!, :res, :(Symbol($(string(sol_field,"_sols")))) , Expr(:ref, sol_type))
+
+    if action_kw == :update
+        body = quote
+            b = $bound
+            is_inc_sol = $call_to_update
+            if is_inc_sol
+                $body
+            end
+        end
+    elseif action_kw == :add
+        body = quote
+            $body
+            b = $bound
+            $call_to_update
+        end
     end
-    pb = PrimalBound{S}(getvalue(solution))
-    inc_sol = force ? true : update_ip_primal_bound!(res.incumbents, pb)
-    if inc_sol && length(res.ip_primal_sols) > 0
-        res.ip_primal_sols[1] = solution
-    elseif length(res.ip_primal_sols) == 0
-        push!(res.ip_primal_sols, solution)
+    code = quote
+        function $method_name(res::OptimizationResult{F,S}, sol::$(sol_type)) where {F,S}
+            $max_len == 0 && return
+            if $field === nothing
+                $define_array
+            end
+            $(body)
+            return
+        end
     end
-    #sort!(res.ip_primal_sols; by = x -> valueinminsense(PrimalBound{S}(getvalue(x))))
-    return
+    return esc(code)
 end
 
-function add_lp_primal_sol!(res::OptimizationResult{F,S}, solution::PrimalSolution{F}, force = false) where {F,S}
-    if res.lp_primal_sols === nothing
-        res.lp_primal_sols = PrimalSolution{F}[]
-    end
-    pb = PrimalBound{S}(getvalue(solution))
-    inc_sol = force ? true : update_lp_primal_bound!(res.incumbents, pb)
-    if inc_sol && length(res.lp_primal_sols) > 0
-        res.lp_primal_sols[1] = solution
-    elseif length(res.lp_primal_sols) == 0
-        push!(res.lp_primal_sols, solution)
-    end
-    #sort!(res.lp_primal_sols; by = x -> valueinminsense(PrimalBound{S}(getvalue(x))))
-    return
-end
+@gen_new_sol_method update, ip_primal, Primal
+@gen_new_sol_method add, ip_primal, Primal
+@gen_new_sol_method set, ip_primal, Primal
 
-function add_lp_dual_sol!(res::OptimizationResult{F,S}, solution::DualSolution{F}, force = false) where {F,S}
-    if res.lp_dual_sols === nothing
-        res.lp_dual_sols = DualSolution{F}[]
-    end
-    db = DualBound{S}(getvalue(solution))
-    inc_sol = force ? true : update_lp_dual_bound!(res.incumbents, db)
-    if inc_sol && length(res.lp_dual_sols) > 0
-        res.lp_dual_sols[1] = solution
-    elseif length(res.lp_dual_sols) == 0
-        push!(res.lp_dual_sols, solution)
-    end
-    #sort!(res.lp_dual_sols; by = x -> valueinminsense(PrimalBound{S}(getvalue(x))))
-    return
-end
+@gen_new_sol_method update, lp_primal, Primal
+@gen_new_sol_method add, lp_primal, Primal
+@gen_new_sol_method set, lp_primal, Primal
+
+@gen_new_sol_method update, lp_dual, Dual
+@gen_new_sol_method add, lp_dual, Dual
+@gen_new_sol_method set, lp_dual, Dual
+
 
 function Base.print(io::IO, form::AbstractFormulation, res::OptimizationResult)
     println(io, "┌ Optimization result ")
